--- a/go.mod
+++ b/go.mod
@@ -4,5 +4,5 @@ go 1.13
 
 require (
 	github.com/arbovm/levenshtein v0.0.0-20160628152529-48b4e1c0c4d0
-	github.com/dgryski/trifles v0.0.0-20190318185328-a8d75aae118c
+	github.com/dgryski/trifles v0.0.0-20200323201526-dd97f9abfb48
 )