--- a/go.mod
+++ b/go.mod
@@ -1,9 +1,8 @@
-module github.com/psadac/levenshtein
+module github.com/agnivade/levenshtein
 
 go 1.13
 
 require (
-	github.com/agnivade/levenshtein v1.0.3
 	github.com/arbovm/levenshtein v0.0.0-20160628152529-48b4e1c0c4d0
 	github.com/dgryski/trifles v0.0.0-20200323201526-dd97f9abfb48
 )