--- a/levenshtein.go
+++ b/levenshtein.go
@@ -54,9 +54,8 @@ func ComputeDistance(a, b string) int {
 		prev := uint16(i)
 		var current uint16
 		for j := 1; j <= lenS1; j++ {
-			if s2[i-1] == s1[j-1] {
-				current = x[j-1] // match
-			} else {
+			current = x[j-1] // match
+			if s2[i-1] != s1[j-1] {
 				current = min(min(x[j-1]+1, prev+1), x[j]+1)
 			}
 			x[j-1] = prev