I am not sure that the removal of the cache initialisation (last commit) works in every situation but all the tests are ok.