@psadac 2020-05-11
You are fully right to be cautious, it does not work. 
If you had a line `{"ab", "ba", 2},` in your TestSanity() test function, it would have generated an error.