@agnivade 2020-05-12
Haha nice work with the loop. Shall we just keep the if condition optimization then ? And also add the ab/ba (abba !) test case while at it ?