@psadac 2020-05-13
The Abba sanity case has been added ;-)
Yes, the condition optimization is still relevant. Comparing your latest commit 7df6e981 and "Declare "current" variable in the inner loop" f1a1c597 will give the following benchmarks :

```
name              old time/op    new time/op    delta
Simple/ASCII-4       371ns ±11%     306ns ± 3%  -17.41%  (p=0.000 n=19+19)
Simple/French-4      672ns ± 4%     575ns ± 5%  -14.52%  (p=0.000 n=20+19)
Simple/Nordic-4     1.39µs ± 4%    1.08µs ± 4%  -22.37%  (p=0.000 n=19+20)
Simple/Tibetan-4    1.07µs ± 5%    0.98µs ± 4%   -8.75%  (p=0.000 n=20+20)
[Geo mean]           781ns          657ns       -15.91%
```
For the relevant part the assembly code generated show there's one less jump and one less move, perhaps it can help to explain such a difference ?

code before :

```assembly
	0x0174 00372 (levenshtein.go:57)	MOVL	-4(SI)(R8*4), R10
	0x0179 00377 (levenshtein.go:57)	MOVL	-4(BX)(R9*4), R11
	0x017e 00382 (levenshtein.go:57)	CMPL	R10, R11
	0x0181 00385 (levenshtein.go:57)	JNE	395
	0x0183 00387 (levenshtein.go:58)	MOVWLZX	-2(AX)(R9*2), R10
	0x0189 00393 (levenshtein.go:58)	JMP	355
	0x018b 00395 (levenshtein.go:60)	MOVWLZX	-2(AX)(R9*2), R11
```
and after :

```assembly
	0x0174 00372 (levenshtein.go:56)	MOVWLZX	-2(AX)(R9*2), R10
	0x017a 00378 (levenshtein.go:57)	MOVL	-4(CX)(R8*4), R11
	0x017f 00383 (levenshtein.go:57)	MOVL	-4(SI)(R9*4), R12
	0x0184 00388 (levenshtein.go:57)	CMPL	R11, R12
	0x0187 00391 (levenshtein.go:57)	JEQ	355
```