@agnivade 2020-05-13
Sounds good. I'll take a deeper look at the assembly when I get some time. Let's revert the go.mod changes and the `TestCacheWithoutInit` function ? That seems to have served its purpose.