@agnivade 2020-05-24
Ok, it looks like the branch predictor was somehow not performing optimally. So removing a branch increases the instruction throughput.

Here is the perf stat comparisons

Performance counter stats for './levenshtein.test -test.bench=Simple/ASCII' (5 runs):

PR
```
          1,512.35 msec task-clock                #    1.043 CPUs utilized            ( +-  0.49% )
             3,232      context-switches          #    0.002 M/sec                    ( +-  1.45% )
                45      cpu-migrations            #    0.030 K/sec                    ( +- 14.49% )
             1,614      page-faults               #    0.001 M/sec                    ( +-  2.18% )
    5,39,24,34,849      cycles                    #    3.566 GHz                      ( +-  1.06% )  (55.95%)
   18,33,52,84,672      instructions              #    3.40  insn per cycle           ( +-  1.07% )  (70.73%) -- more instructions
    3,55,67,93,143      branches                  # 2351.836 M/sec                    ( +-  1.07% )  (71.35%)
         16,48,351      branch-misses             #    0.05% of all branches          ( +-  2.11% )  (72.19%) -- fewer branch misses
    3,66,83,33,827      L1-dcache-loads           # 2425.589 M/sec                    ( +-  1.23% )  (72.61%)
         59,57,345      L1-dcache-load-misses     #    0.16% of all L1-dcache hits    ( +-  1.73% )  (71.86%)
   <not supported>      L1-icache-loads                                             
         75,76,911      L1-icache-load-misses                                         ( +-  5.48% )  (56.03%)

           1.45003 +- 0.00774 seconds time elapsed  ( +-  0.53% )
```

Master
```
          1,559.70 msec task-clock                #    1.040 CPUs utilized            ( +-  0.29% )
             3,212      context-switches          #    0.002 M/sec                    ( +-  1.68% )
                52      cpu-migrations            #    0.033 K/sec                    ( +- 13.46% )
             1,632      page-faults               #    0.001 M/sec                    ( +-  1.59% )
    5,58,00,68,480      cycles                    #    3.578 GHz                      ( +-  0.33% )  (56.06%)
   16,75,68,62,901      instructions              #    3.00  insn per cycle           ( +-  0.30% )  (70.89%) 
    3,31,79,64,986      branches                  # 2127.308 M/sec                    ( +-  0.28% )  (71.55%)
         29,02,301      branch-misses             #    0.09% of all branches          ( +-  3.86% )  (72.31%)
    3,35,48,26,049      L1-dcache-loads           # 2150.941 M/sec                    ( +-  0.53% )  (72.50%)
         55,48,670      L1-dcache-load-misses     #    0.17% of all L1-dcache hits    ( +-  2.19% )  (71.62%)
   <not supported>      L1-icache-loads                                             
         83,97,670      L1-icache-load-misses                                         ( +-  7.65% )  (55.95%)

           1.49978 +- 0.00387 seconds time elapsed  ( +-  0.26% )
```

The PR increases throughput to 3.4 instructions/cycle from 3 instructions/cycle. I'd guess the reason to be the reduction in branch misses from 0.09% to 0.05%.